+++ runtime/src/commonMain/kotlin/pbandk/impl/Sizer.kt
@@ -2,22 +2,55 @@ package pbandk.impl
 
 import pbandk.*
 
+private class CodePointIterator(private val s: String) : Iterator<Int> {
+    var pos = 0
+
+    override fun hasNext(): Boolean = pos < s.length
+
+    override fun next(): Int {
+        if (pos >= s.length) throw NoSuchElementException()
+
+        val v = s[pos++]
+        if (v.isHighSurrogate() && pos < s.length) {
+            val l = s[pos]
+            if (l.isLowSurrogate()) {
+                pos++
+                return 0x10000 + (v - 0xD800).toInt() * 0x400 + (l - 0xDC00).toInt()
+            }
+        }
+        return v.toInt() and 0xffff
+    }
+}
+
+private class CodePointIterable(private val s: String) : Iterable<Int> {
+    override fun iterator(): Iterator<Int> = CodePointIterator(s)
+}
+
+fun utf8Len(value: String) = CodePointIterable(value).sumBy {
+    when (it) {
+        in 0..0x7f -> 1
+        in 0x80..0x7ff -> 2
+        in 0x800..0xffff -> 3
+        else -> 4
+    }
+}
+
 open class SizerImpl : Sizer {
     override fun stringSize(value: String): Int {
-        val len = UtilImpl.utf8Len(value)
+        val len = utf8Len(value)
         return uInt32Size(len) + len
     }
 
     override fun enumSize(value: Message.Enum) = int32Size(value.value)
     override fun messageSize(value: Message<*>) = uInt32Size(value.protoSize) + value.protoSize
     override fun <T> packedRepeatedSize(list: List<T>, sizeFn: (T) -> Int) =
-            if (list is ListWithSize && list.protoSize != null) list.protoSize + uInt32Size(list.protoSize)
-            else list.sumBy(sizeFn).let { it + uInt32Size(it) }
+        if (list is ListWithSize && list.protoSize != null) list.protoSize + uInt32Size(list.protoSize)
+        else list.sumBy(sizeFn).let { it + uInt32Size(it) }
 
     override fun <K, V, T : Message<T>> mapSize(
-            fieldNumber: Int,
-            map: Map<K, V>,
-            createEntry: (K, V, Map<Int, pbandk.UnknownField>) -> T
+        fieldNumber: Int,
+        map: Map<K, V>,
+        createEntry: (K, V, Map<Int, pbandk.UnknownField>) -> T
     ) = tagSize(fieldNumber).let { tagSize ->
         map.entries.sumBy { e ->
             val msg = e as? Message<*> ?: createEntry(e.key, e.value, emptyMap())
