+++ runtime/src/nativeMain/kotlin/pbandk/Util.kt
@@ -1,7 +1,14 @@
 package pbandk
 
+import pbandk.impl.AbstractUtil
+
 actual object SizerImpl: pbandk.impl.SizerImpl(), Sizer
-actual typealias UtilImpl = pbandk.impl.UtilImpl
+
+@OptIn(ExperimentalStdlibApi::class)
+actual object UtilImpl : AbstractUtil(), Util {
+        override fun stringToUtf8(str: String): ByteArray = str.encodeToByteArray()
+        override fun utf8ToString(bytes: ByteArray): String = bytes.decodeToString()
+}
 
 internal actual fun unmarshallerByteArray(arr: ByteArray): Unmarshaller =
         pbandk.impl.Unmarshaller.fromByteArray(arr)
